Updated scripts and docker

+++ homeserver/scripts/wireguard_peer.sh
@@ -1,45 +1,66 @@
 #!/bin/bash
 
 # Collect peer name and its wireguard IP from user
-read -p "Enter Name for the Peer: " peer
-[ -z $peer ] && echo -e "\nPeer Name Unavailable...Exiting." && exit
-echo ""
-read -p "Enter Wireguard IP for the Peer: " peerip
-[ -z $peerip ] && echo -e "\nPeer IP Unavailable...Exiting." && exit
+read -p "\nEnter Name for the Peer: " PEER
+[[ -z ${PEER} ]] && echo -e "\nPeer Name Unavailable...Exiting." && exit
+read -p "\nEnter Wireguard IP for the Peer: " PEERIP
+[[ -z ${PEERIP} ]] && echo -e "\nPeer IP Unavailable...Exiting." && exit
 
 # Set basic parameters
 USER=$(whoami)
-WGDIR=/home/$USER/wireguard
-PEERDIR=/home/$USER/wireguard/$peer
-SERVERDIR=/etc/wireguard
-SERVERIP="$(sudo cat $SERVERDIR/wg0.conf | grep "Address" | cut -d " " -f 3 | cut -d "/" -f 1)"
-LISTENPORT="$(sudo cat $SERVERDIR/wg0.conf | grep "ListenPort" | cut -d " " -f 3)"
-# If you are using a domain name or dynamic dns, update below value of WANIP
-WANIP="$(wget -q -O - https://ifconfig.io/ip)"
-#WANIP="mydomain.example.com"
+WGDIR="/home/${USER}/wireguard"
+PEERDIR="/home/${USER}/wireguard/${PEER}"
+SERVERDIR="/etc/wireguard"
+SERVERIP="$(sudo cat ${SERVERDIR}/wg0.conf | grep "Address" | cut -d " " -f 3 | cut -d "/" -f 1)"
+LISTENPORT="$(sudo cat ${SERVERDIR}/wg0.conf | grep "ListenPort" | cut -d " " -f 3)"
+
+# Identify server domain or public IP
+VALIDATE="^([a-zA-Z0-9][a-zA-Z0-9-]{0,61}[a-zA-Z0-9]\.)+[a-zA-Z]{2,}$"
+read -p "\nIs your server accessible via a domain? (y/N): " DOMAINIP
+case ${DOMAINIP} in
+    [yY]* ) read -p "\nEnter your server domain (myserver.example.com): " WANIP
+            if [[ -z ${WANIP} ]]; then
+                echo -e "\nDomain Name Unavailable...Using Public IP."
+                WANIP="$(wget -q -O - https://ifconfig.io/ip)"
+            else
+                if [[ ${WANIP} =~ ${VALIDATE} ]]; then
+                    echo -e "\nValid domain name. Using ${WANIP} to access Wireguard Server."
+                else
+                    echo -e "\nInvalid Domain Name...Using Public IP."
+                    WANIP="$(wget -q -O - https://ifconfig.io/ip)"
+                fi
+            fi
+        ;;
+    [nN]* ) echo -e "\nDomain Name Unavailable...Using Public IP."
+            WANIP="$(wget -q -O - https://ifconfig.io/ip)"
+        ;;
+    * ) echo "Invalid response...Using server Public IP."
+            WANIP="$(wget -q -O - https://ifconfig.io/ip)"
+        ;;
+esac
 
 # Check for peer directory or generate
 umask 077
-[ -d $WGDIR ] || mkdir $WGDIR
-[ -d $PEERDIR ] || mkdir $PEERDIR
+[[ -d ${WGDIR} ]] || mkdir ${WGDIR}
+[[ -d ${PEERDIR} ]] || mkdir ${PEERDIR}
 
 # Generate peer keys
-wg genkey > $PEERDIR/$peer.key
-wg pubkey < $PEERDIR/$peer.key > $PEERDIR/$peer.pub
-wg genpsk > $PEERDIR/$peer.psk
+wg genkey > ${PEERDIR}/${PEER}.key
+wg pubkey < ${PEERDIR}/${PEER}.key > ${PEERDIR}/${PEER}.pub
+wg genpsk > ${PEERDIR}/${PEER}.psk
 
 # Generate peer configuration file
 echo "\
 [Interface]
-Address = $peerip/32
-PrivateKey = $(cat $PEERDIR/$peer.key)
-DNS = $SERVERIP
+Address = ${PEERIP}/32
+PrivateKey = $(cat ${PEERDIR}/${PEER}.key)
+DNS = ${SERVERIP}
 
 [Peer]
-Endpoint = $WANIP:$LISTENPORT
+Endpoint = ${WANIP}:${LISTENPORT}
 AllowedIPs = 0.0.0.0/0
-PublicKey = $(sudo cat $SERVERDIR/server.pub)
-PresharedKey = $(cat $PEERDIR/$peer.psk)" > $PEERDIR/$peer.conf
+PublicKey = $(sudo cat ${SERVERDIR}/server.pub)
+PresharedKey = $(cat ${PEERDIR}/${PEER}.psk)" > ${PEERDIR}/${PEER}.conf
 
 # Shutdown wireguard server to save new configuration
 echo -e "\nShutting down wireguard server..."
@@ -47,30 +68,30 @@ sudo wg-quick down wg0
 
 # Backup current server configuration
 echo -e "\nBacking up current server configuration..."
-sudo cp $SERVERDIR/wg0.conf $SERVERDIR/wg0.conf.bak
+sudo cp ${SERVERDIR}/wg0.conf ${SERVERDIR}/wg0.conf.bak
 
 # Update server configuration
 echo "\
 
 [Peer]
-AllowedIPs = $peerip/32
-PublicKey = $(cat $PEERDIR/$peer.pub)
-PresharedKey = $(cat $PEERDIR/$peer.psk)" | sudo tee -a $SERVERDIR/wg0.conf > /dev/null
+AllowedIPs = ${PEERIP}/32
+PublicKey = $(cat ${PEERDIR}/${PEER}.pub)
+PresharedKey = $(cat ${PEERDIR}/${PEER}.psk)" | sudo tee -a ${SERVERDIR}/wg0.conf > /dev/null
 
 echo -e "\nWireguard profile for $peer generated."
 echo -e "\nRestarting wireguard server..."
 sudo wg-quick up wg0
-echo ""
 
 # Generate QR code of peer.conf file
-if ! [ -x "$(command -v qrencode)" ]; then
-    echo "qrencode is not installed. QR code generation skipped." >&2
+if ! [[ -x "$(command -v qrencode)" ]]; then
+    echo "\nqrencode is not installed. QR code generation skipped." >&2
+    echo "\nWireguard Peer Profile Generation Completed."
     exit
 fi
 
-read -p "Do you want to generate peer configuration QR code? (y/n): " yn
-case $yn in
-    [yY]* ) qrencode -t ansiutf8 -r "$PEERDIR/$peer.conf"
+read -p "Do you want to generate peer configuration QR code? (y/n): " YN
+case ${YN} in
+    [yY]* ) qrencode -t ansiutf8 -r "${PEERDIR}/${PEER}.conf"
         ;;
     [nN]* ) echo "QR code generation cancelled...Exiting."
         ;;
